requirements.txt
requests==2.32.5
fastapi==0.115.14
pydantic==2.10.6
python-dotenv==1.1.0